Updated project structure and .gitignore, Fixed FCCAPIUtilTest, took out util.logging
