+++ .gitignore
@@ -16,6 +16,7 @@ target/
 .sts4-cache
 
 ### IntelliJ IDEA ###
+/.idea/
 .idea
 *.iws
 *.iml
