+++ README.md
@@ -26,7 +26,7 @@ __5$ * 2 = 10$__
     результат с учетом указанного курса обмена;
 2) Выполнять умножение величин в валюте (стоимость одной акции) на число
    (количество акций), результатом этой операции должна быть величина в
-    вылюте.
+    валюте.
 3) Сделать переменную amount закрытым членом класса
 4) Побочные эффекты(side effect) в классе Dollar
 5) Округление денежных величин
@@ -43,7 +43,7 @@ __5$ * 2 = 10$__
 2) Запустить все тесты, при этом обнаружить, что что-то не срабатывает.</br>
 3) Внести небольшое изменение.</br>
 4) Снова запустить тесты и убедиться, что все они успешно выполняются.</br>
-5) Устранить доблирование с помощью рефакторинга</br>
+5) Устранить дублирование с помощью рефакторинга</br>
 
 Пример:</br>
