+++ bank/src/main/java/ru/ithub/annotation/ApiModule.java
