+++ bank/src/main/java/ru/ithub/converter/CheckConverter.java
