+++ bank/src/main/java/ru/ithub/converter/impl/CheckConverterImpl.java
