+++ bank/src/main/java/ru/ithub/currency/Currency.java
