+++ bank/src/main/java/ru/ithub/currency/CurrencyPair.java
