+++ bank/src/main/java/ru/ithub/exception/AlreadyExistException.java
