+++ bank/src/main/java/ru/ithub/exception/ApiRequestException.java
