+++ bank/src/test/java/ru/ithub/entity/CheckTest.java
