+++ bank/src/test/java/ru/ithub/factory/impl/CheckFactoryTest.java
