+++ logging/src/main/java/ru/fadesml/libs/AutoLoggedComponent.java
@@ -1,0 +1,7 @@
+package ru.fadesml.libs;
+
+import java.util.logging.Logger;
+
+public interface AutoLoggedComponent {
+    Logger getLogger();
+}
