+++ logging/src/main/java/ru/fadesml/libs/LoggingType.java
@@ -1,0 +1,6 @@
+package ru.fadesml.libs;
+
+public enum LoggingType {
+    ALL,
+    MARKED
+}
