+++ pom.xml
@@ -4,19 +4,29 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>junit-tdd-ithub</artifactId>
+    <groupId>ru.ithub</groupId>
+    <artifactId>parent</artifactId>
     <version>1.0-SNAPSHOT</version>
-    <packaging>jar</packaging>
+    <packaging>pom</packaging>
 
     <name>tdd-by-ithub</name>
     <description>Test Driven Development By ithub.ru</description>
+
+    <modules>
+        <module>logging</module>
+        <module>bank</module>
+    </modules>
 
     <developers>
         <developer>
             <id>tns</id>
             <name>Temnyakov Nikolay</name>
             <email>temnjakovn@gmail.com</email>
+        </developer>
+        <developer>
+            <id>fadesml</id>
+            <name>Andrew Buniak</name>
+            <email>bakite.buniak@gmail.com</email>
         </developer>
     </developers>
 
@@ -26,31 +36,9 @@
         <java.version>11</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <junit-platform.version>5.7.2</junit-platform.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
     </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-engine</artifactId>
-            <version>${junit-platform.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-api</artifactId>
-            <version>${junit-platform.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-            <version>3.6.28</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
 
     <build>
         <plugins>
@@ -71,5 +59,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
