--- src/main/java/ru/ithub/annotation/AutoLogged.java
@@ -1,11 +1,0 @@
-package ru.ithub.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface AutoLogged {
-}
