--- src/main/java/ru/ithub/annotation/Parameter.java
@@ -1,12 +1,0 @@
-package ru.ithub.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface Parameter {
-    String name();
-}
