--- src/main/java/ru/ithub/util/logging/AutoLoggedComponent.java
@@ -1,7 +1,0 @@
-package ru.ithub.util.logging;
-
-import java.util.logging.Logger;
-
-public interface AutoLoggedComponent {
-    Logger getLogger();
-}
