--- src/main/java/ru/ithub/util/logging/LoggingType.java
@@ -1,6 +1,0 @@
-package ru.ithub.util.logging;
-
-public enum LoggingType {
-    ALL,
-    MARKED
-}
