--- src/test/java/package-info.java
@@ -1,7 +1,0 @@
-package ithub.ru;
-/*
- * @created 29/08/2021 - 18:23
- * @project IntelliJ IDEA
- * @author Temnyakov Nikolay, Andrew "Fadesml" Bunyak
- */
-
