--- src/test/java/ru/ithub/util/ErrorHttpResponse.java
@@ -1,57 +1,0 @@
-package ru.ithub.util;
-
-import javax.net.ssl.SSLSession;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpHeaders;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.util.Optional;
-
-/**
- * Класс заплатки HttpResponse - эмитирует ошибочный ответ при запросе к api
- *
- * @see java.net.http.HttpResponse
- * @see FCCAPIUtil
- */
-public class ErrorHttpResponse implements HttpResponse {
-    @Override
-    public int statusCode() {
-        return 400;
-    }
-
-    @Override
-    public HttpRequest request() {
-        return null;
-    }
-
-    @Override
-    public Optional<HttpResponse> previousResponse() {
-        return Optional.empty();
-    }
-
-    @Override
-    public HttpHeaders headers() {
-        return null;
-    }
-
-    @Override
-    public Object body() {
-        return null;
-    }
-
-    @Override
-    public Optional<SSLSession> sslSession() {
-        return Optional.empty();
-    }
-
-    @Override
-    public URI uri() {
-        return null;
-    }
-
-    @Override
-    public HttpClient.Version version() {
-        return null;
-    }
-}
